Fix: Moved application data model to correct folder
